# doc/api.yaml
...
openapi: 3.0.0 
info:  
  title: WasaPhoto 
  description:  |  
    This API allows you to manage the 
    data available in the project 
  version: 0.0.1 
servers:
  - url: http://localhost:3000
paths:  
...
  /users/{userId}/followUser/{userId2}: 
    parameters:  
      - $ref: "#/components/parameters/userId" 
      - $ref: "#/components/parameters/userId2"     

    put: 
          tags: ["followUser"] 
          summary: Follow another user 
          description: | 
            If the user identified by userId or userId2 does not exist, error is returned. 
            If the user identified by userId2 exists, user identified by userId follow user identified by userId2 if user2 hasn't banned him. 
          operationId: followUser 
          security: 
            - BearerAuth: [] 

          responses:  
            '200': 
              description: Now you follow the other user 
              content: 
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/scUserName' 
                  example:
                    username: "Luigi11" 

            '400': { $ref: "#/components/responses/UserIdNotFound" }                  
            '403': {$ref: "#/components/responses/UserIdBanned"}
            '404': {$ref: "#/components/responses/UserId2NotFound"} 
            '405': {$ref: "#/components/responses/userId2Banned"}
            '500': {$ref: "#/components/responses/ErrorServerExecution"}  
            '502': {$ref: "#/components/responses/FailedAutentication"}  
            '503': {$ref: "#/components/responses/NotAllowed"}  

    delete:
          tags: ["unfollowUser"] 
          summary: Unfollow another user 
          description: | 
            If the user identified by userId or userId2 does not exist or user doesn't follow userId2, error is returned. 
            If user follow userId2, user identified by userId unfollow user identified by userId2. 
          operationId: unfollowUser 
          security: 
            - BearerAuth: [] 

          responses:  
            '200': 
              description: Now you don't follow the other user 
              content: 
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/scUserName' 
                  example:
                    username: "Luigi11" 

            '400': { $ref: "#/components/responses/UserIdNotFound" }                  
            '403': {$ref: "#/components/responses/UserIdBanned"}   
            '402': {$ref: "#/components/responses/UnfollowedUser"}    
            '404': {$ref: "#/components/responses/UserId2NotFound"}
            '405': {$ref: "#/components/responses/userId2Banned"}   
            '500': {$ref: "#/components/responses/ErrorServerExecution"}  
            '502': {$ref: "#/components/responses/FailedAutentication"}  
...
  /users/{userId}/photo:
    parameters:
      - $ref: '#/components/parameters/userId'
    post:
          tags: ["uploadPhoto"] 
          summary:  Upload new photo 
          description: | 
            If the user identified by userId  doesn't exist, error is returned. 
            If the user identified by userId exist, the Photo is uploaded.
          operationId: uploadPhoto 
          security: 
            - BearerAuth: [] 
          requestBody:
           description: The binary image file to upload
           content:
            image/*:
              schema:
                  $ref: "#/components/schemas/image"
          responses:
            '201': 
              description: Photo uploaded
              content: 
                application/json: 
                  schema: 
                    $ref: '#/components/schemas/errorSchema' 
                  example:
                    result: "Done" 

            '400': { $ref: "#/components/responses/UserIdNotFound" }                       
            '410': { $ref: "#/components/responses/InvalidFormat" }   
            '500': {$ref: "#/components/responses/ErrorServerExecution"}  
            '502': {$ref: "#/components/responses/FailedAutentication"}  

          
    get:
      tags: ["getUserPhotos"]
      summary: Returns user photos
      description: |
        If the user identified by userId  doesn't exist, error is returned. 
        If the user identified by userId exist, returns the list of photos uploaded by a user
      operationId: getUserPhotos
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Returns the user photos list
          content:
            application/json:
              schema:
                type: array 
                minItems: 0
                maxItems: 100
                description:  Various posts of the user 
                items:
                  $ref: '#/components/schemas/userPhotoStream'
              example:
                - photoId: 12
                  userId: 3
                  name: "Mario_1"
                  likes: 109
                  comments: 124
                  upladTime: "2023-02-12T12:00:00Z"
                - photoId: 52
                  userId: 3
                  name: "Mario_1"
                  likes: 10
                  comments: 18
                  upladTime: "2023-03-18T12:00:00Z"

        '400': { $ref: "#/components/responses/UserIdNotFound" }    
        '403': {$ref: "#/components/responses/UserIdBanned"}
        '404': {$ref: "#/components/responses/UserId2NotFound"}
        '405': {$ref: "#/components/responses/userId2Banned"}     
        '500': {$ref: "#/components/responses/ErrorServerExecution"}    
...
  /users:
    get:
      tags: ["getUserId"]
      summary: Get userId from his username
      description:  |
        If user identified by userId in authentication doesn't exist or user identified by the username passed doesn't exist, error is returned.
        If user identified by userId in authentication exist and user identified by the username passed exist, the userId of the user is ruterned.
      operationId: getUserId
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: query
          schema:
            $ref: "#/components/schemas/username"
          required: true
          description: Username of the user to search 
      responses:
          '201': 
            description: User log-in action successful 
            content: 
              application/json: 
                schema: 
                    $ref: "#/components/schemas/userId"
                example:
                  userId: 111112
          '404': { $ref: "#/components/responses/UsernameNotFound" }    
          '400': { $ref: "#/components/responses/UserIdNotFound" }    
          '500': {$ref: "#/components/responses/ErrorServerExecution"} 
          '411': { $ref: "#/components/responses/UsernameNotValid" }
...
components:  
  responses:
...
    UsernameNotValid:
     description: Username not valid
    FailedAutentication:
      description: Failed authentication
    UsernameNotFound:
      description: Searched username not found
    UserIdNotFound:
      description: User id not found
...
    UserIdBanned:
      description: User banned by User2
    userId2Banned:
      description: User banned User2
    UserId2NotFound:
      description: Userid2 not found    
    ErrorServerExecution: 
      description: Error server execution     
    UnfollowedUser:
      description: You don't follow userId2 
...
    InvalidFormat:
      description: Invalid file format
    NotAllowed:
      description: "Operation not allowed"
  schemas: 
    username: 
      type: string 
      description: User's UserName 
      example: "Franchino_12"
      pattern: '^[a-zA-Z0-9_.-]*$' 
      minLength: 6 
      maxLength: 16  
    userId:
      type: object 
      description: User log-in action successful 
      properties: 
        identifier: 
          $ref: '#/components/schemas/identifier'
    identifier: 
      type: integer
      description: User Id 
      example: 1 
...
    photoId:
      type: integer
      description: Photo identifier
      example: 5
    likeNumber: 
      type: integer
      description: Number of likes of the post 
      example: 987
    commentNumber:
      type: integer
      description: Number of comment of the post  
      example: 10000
...
    uploadTime:
      type: string
      format: date-time
      description: Upload time
      minLength: 20
      maxLength: 25
      example: "2022-01-13T12:00:00Z"
    image:
      description: The photo file
      type: string
      format: binary
      minLength: 1
      maxLength: 10485760 
      pattern: "((.|\n)*)" 
...
    userPhotoStream:
      type: object
      properties:
        photoId:
          $ref: '#/components/schemas/photoId'
        likes:
          $ref: '#/components/schemas/likeNumber'
        comments: 
          $ref: '#/components/schemas/commentNumber'
        uploadTime:
          $ref: '#/components/schemas/uploadTime'
        userId: 
          $ref: '#/components/schemas/identifier'
        name: 
          $ref: '#/components/schemas/username'
...
    errorSchema: 
      description: Error example
      type: object 
      properties: 
        Message: 
          $ref: "#/components/schemas/errorDescription" 
    errorDescription: 
      type: string  
      description: Status
      example: Success 
      pattern: '^[a-zA-Z0-9]+$' 
      minLength: 4
      maxLength: 50 
...
    scUserName:
      type: object
      description:  "Describe userName" 
      properties:
        username:
          $ref: "#/components/schemas/username"  
...
  parameters:  
    userId: 
      in: path 
      name: userId 
      required: true 
      description: User Id 
      schema: 
        type: integer 
        description: "User Id" 
        example: 86 
    userId2: 
      in: path 
      name: userId2 
      required: true 
      description: Second userId
      schema: 
        type: integer 
        description: "Second userId" 
        example: 100 
...
  securitySchemes: 
    BearerAuth: 
      description: login only as your user id 
      type: http 
      scheme: bearer 